requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
SQLAlchemy==2.0.21
PyMySQL==1.1.0
email-validator==2.0.0
smtplib
logging